07-04-23_20-13

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->id();
             $table->string('username')->unique();
             $table->string('email')->unique();
-            $table->integer('rating');
+            $table->decimal('rating',2,1);
             $table->string('password');
             $table->string('foto')->unsinged()->nullable();
             $table->enum('pangkat', ['awam','otakers','admin']);
